py_slr_algorithm_rel1/requirements.txt
numpy
pandas        	
Pillow   
scikit-learn 
streamlit
pyinstaller

